Move <CheckboxGroup /> and <RadioButtonGroup /> to the root
