+++ lib/CheckboxGroup.js
