+++ lib/RadioButtonGroup.js
@@ -5,7 +5,7 @@
 
 var React         = require('react');
 var PropTypes     = React.PropTypes;
-var emptyFunction = require('../emptyFunction');
+var emptyFunction = require('./emptyFunction');
 
 function renderEmptyOption(props, onChange) {
   return (
