+++ lib/__tests__/CheckboxGroup-test.js
