+++ lib/index.js
@@ -3,22 +3,22 @@
  */
 'use strict';
 
-var Form                    = require('./Form');
+var Form              = require('./Form');
 
-var Field                   = require('./Field');
-var Fieldset                = require('./Fieldset');
-var RepeatingFieldset       = require('./RepeatingFieldset');
-var Element                 = require('./Element');
-var Input                   = require('./Input');
+var Field             = require('./Field');
+var Fieldset          = require('./Fieldset');
+var RepeatingFieldset = require('./RepeatingFieldset');
+var Element           = require('./Element');
+var Input             = require('./Input');
 
-var Message                 = require('./Message');
-var Label                   = require('./Label');
+var Message           = require('./Message');
+var Label             = require('./Label');
 
-var PropTypes               = require('./PropTypes');
+var PropTypes         = require('./PropTypes');
 
-var messages                = require('./messages');
-var types                   = require('./types');
-var schema                  = require('./schema');
+var messages          = require('./messages');
+var types             = require('./types');
+var schema            = require('./schema');
 
 module.exports = {
   Form,
