--- lib/input/index.js
@@ -1,8 +1,0 @@
-'use strict';
-/**
- * @jsx React.DOM
- */
-module.exports = {
-  CheckboxGroup: require('./CheckboxGroup'),
-  RadioButtonGroup: require('./RadioButtonGroup')
-};
