+++ lib/types.js
@@ -19,7 +19,7 @@ class Type {
 }
 
 function idSerialize(value) {
-  return value === null ? '' : value;
+  return value == null ? '' : value;
 }
 
 function idDeserialize(value) {
@@ -47,7 +47,7 @@ var isDateRe = /^\d\d\d\d-\d\d-\d\d$/;
 
 var date = new Type('date',
   function(value) {
-    if (value === null) {
+    if (value == null) {
       return '';
     }
     var year = value.getFullYear();
